Added stand, anchors and sockets and build onto headset
