+++ Assets/Materials/Solid Colours/Purple.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 6db072edb853914439f1cfffcc11d644
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 2100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
