+++ Assets/Prefabs/Experiments/Extraction.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 3de81e9947b6a2640bfad9e384f2d387
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
