+++ Assets/Prefabs/Temp Equipments.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 63d5f9a6bfbc01440a29118ac1aed072
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
