+++ Assets/Prefabs/Temp Equipments/Separatory Funnel.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 873bd6804c8c0e34e927a2e842a06670
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
