+++ Assets/Prefabs/Temp Equipments/Stopper.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 83d382d2ac9e51243a1e67db12814c76
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
