+++ Assets/Scripts/DetectShaking.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 5a33fe4153e3c664ca23c47a1f461c29
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
