+++ Assets/Shaders/LiquidEffect.shadergraph.meta
@@ -1,0 +1,10 @@
+fileFormatVersion: 2
+guid: 6172518a19a57ff4f8adb5d70d9ca211
+ScriptedImporter:
+  internalIDToNameTable: []
+  externalObjects: {}
+  serializedVersion: 2
+  userData: 
+  assetBundleName: 
+  assetBundleVariant: 
+  script: {fileID: 11500000, guid: 625f186215c104763be7675aa2d941aa, type: 3}
