--- Assets/Temporary/LayeredLiquidController.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 7134b3758214bdf4eb5bc7d1ed9c07fd
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
