--- Assets/Temporary/SeparatoryFunnel.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: e1337ad5bcb39b240b2e6f217fec5278
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
