--- Assets/Temporary/TurnTriggerOnPlaced.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: b93363b328be61d449f3284d5b302acd
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
