--- Assets/Temporary/UxrShakeDetector.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 3c3a7a9cbbd5d7e47a0e9da4d6617a52
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
