+++ ProjectSettings/EditorBuildSettings.asset
@@ -16,7 +16,10 @@ EditorBuildSettings:
     guid: 7408f9471348faf4993f6e6ec0b4a54b
   - enabled: 1
     path: Assets/Scenes/Serial Dilation.unity
-    guid: 838cb0fa6e6a3e1498b80abb3c470680
+    guid: 2affa91b419ac9645ae9d4c3979d285f
+  - enabled: 1
+    path: Assets/Scenes/Extraction.unity
+    guid: d2ed134e83084bc4d804ca2c7b7ae48e
   m_configObjects:
     Unity.XR.Oculus.Settings: {fileID: 11400000, guid: 21beed75abf00ec429d2076d001c932c,
       type: 2}
